chore(portal): moved profileService back to disco-portal
