+++ frontend/apps/portal/src/App.vue
@@ -6,7 +6,7 @@
 import {usePageTitle} from '@disclosure-portal/composables/usePageTitle';
 import DHTTPError from '@shared/types/DHTTPError';
 import ErrorDialogConfig from '@shared/types/ErrorDialogConfig';
-import profileService from '@shared/services/profile';
+import profileService from '@disclosure-portal/services/profile';
 import {useAppStore} from '@disclosure-portal/stores/app';
 import {useCustomIdStore} from '@disclosure-portal/stores/customid.store';
 import {useLabelStore} from '@disclosure-portal/stores/label.store';
