+++ frontend/apps/portal/src/router/index.ts
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as Help from '@disclosure-portal/assets/documents/help';
-import {Rights} from '@shared/types/Rights';
-import profileService from '@shared/services/profile';
+import {Rights} from '@disclosure-portal/model/Rights';
+import profileService from '@disclosure-portal/services/profile';
 import {useUserStore} from '@disclosure-portal/stores/user';
 import Dashboard from '@disclosure-portal/views/Dashboard.vue';
 import {createRouter, createWebHashHistory, RouteRecordRaw} from 'vue-router';
