+++ frontend/libs/portal/components/dialog/TermsOfUseDialog.vue
@@ -4,7 +4,7 @@
 
 <script setup lang="ts">
 import {UserRequestDto} from '@shared/types/Users';
-import ProfileServer from '@shared/services/profile';
+import ProfileServer from '@disclosure-portal/services/profile';
 import {useUserStore} from '@disclosure-portal/stores/user';
 import termsOfUseEn from '@shared/assets/documents/terms_of_use/TermsOfUseCurrent.md?raw';
 import termsOfUseDe from '@shared/assets/documents/terms_of_use/TermsOfUseDe.md?raw';
