+++ frontend/libs/portal/components/dialog/project/settings/ProjectSettings.vue
@@ -6,7 +6,7 @@
 import {PolicyLabels} from '@disclosure-portal/constants/policyLabels';
 import {Project, ProjectSettingsModel} from '@disclosure-portal/model/Project';
 import ProjectPostRequest from '@disclosure-portal/model/ProjectPostRequest';
-import {Group} from '@shared/types/Rights';
+import {Group} from '@disclosure-portal/model/Rights';
 import {useAppStore} from '@disclosure-portal/stores/app';
 import {useDialogStore} from '@disclosure-portal/stores/dialog.store';
 import {useIdleStore} from '@shared/stores/idle.store';
