+++ frontend/libs/portal/components/grids/GridPolicyRulesAssignments.vue
@@ -37,7 +37,7 @@
 <script setup lang="ts">
 import License from '@disclosure-portal/model/License';
 import {PolicyRulesAssignmentsDto} from '@disclosure-portal/model/PolicyRule';
-import {Rights} from '@shared/types/Rights';
+import {Rights} from '@disclosure-portal/model/Rights';
 import {default as LicenseService} from '@disclosure-portal/services/license';
 import {useUserStore} from '@disclosure-portal/stores/user';
 import {computed, onBeforeMount, ref} from 'vue';
