+++ frontend/libs/portal/model/ProjectsResponse.ts
@@ -4,7 +4,7 @@
 
 import {CustomId} from '@disclosure-portal/model/CustomId';
 import {ProjectModel} from '@disclosure-portal/model/Project';
-import {Rights} from '@shared/types/Rights';
+import {Rights} from '@disclosure-portal/model/Rights';
 import {SpdxFile, VersionSlimDto} from '@disclosure-portal/model/VersionDetails';
 import {Application} from './Application';
