+++ frontend/libs/portal/model/Rights.ts
