+++ frontend/libs/portal/stores/user.ts
@@ -4,8 +4,8 @@
 
 import INavItem from '@disclosure-portal/model/INavItem';
 import ITile from '@disclosure-portal/model/ITile';
-import SimpleProfileData from '@shared/types/ProfileData';
-import {Rights} from '@shared/types/Rights';
+import SimpleProfileData from '@disclosure-portal/model/ProfileData';
+import {Rights} from '@disclosure-portal/model/Rights';
 import {UserDto} from '@shared/types/Users';
 import {useAppStore} from '@disclosure-portal/stores/app';
 import {defineStore} from 'pinia';
