+++ frontend/libs/portal/views/Home.vue
@@ -3,7 +3,7 @@
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 
 <script setup lang="ts">
-import ProfileService from '@shared/services/profile';
+import ProfileService from '@disclosure-portal/services/profile';
 import {useAppStore} from '@disclosure-portal/stores/app';
 import {useNewsboxStore} from '@disclosure-portal/stores/newsbox.store';
 import {useUserStore} from '@disclosure-portal/stores/user';
