+++ frontend/libs/portal/views/admin/Labels.vue
@@ -7,7 +7,7 @@ import {ConfirmationType, IConfirmationDialogConfig} from '@disclosure-portal/co
 import useDimensions from '@disclosure-portal/composables/useDimensions';
 import Icons from '@disclosure-portal/constants/icons';
 import Label from '@disclosure-portal/model/Label';
-import {Rights} from '@shared/types/Rights';
+import {Rights} from '@disclosure-portal/model/Rights';
 import AdminService from '@disclosure-portal/services/admin';
 import {useUserStore} from '@disclosure-portal/stores/user';
 import {downloadFile} from '@disclosure-portal/utils/download';
