+++ frontend/libs/portal/views/admin/schema/Schemas.vue
@@ -6,7 +6,7 @@
 import useDimensions from '@disclosure-portal/composables/useDimensions';
 import Icons from '@disclosure-portal/constants/icons';
 import Label from '@disclosure-portal/model/Label';
-import {Rights} from '@shared/types/Rights';
+import {Rights} from '@disclosure-portal/model/Rights';
 import SchemaModel from '@disclosure-portal/model/Schema';
 import AdminService from '@disclosure-portal/services/admin';
 import {useUserStore} from '@disclosure-portal/stores/user';
